parse in hook and add comments
